requirements.txt
# requirements.txt
pandas>=1.5.0
numpy>=1.24.0
faker>=20.1.0
google-cloud-bigquery>=3.13.0
dbt-core>=1.6.0
dbt-bigquery>=1.6.0
apache-airflow>=2.7.0
